+++ NEWS.md
@@ -1,5 +1,7 @@
 # cSEM 0.1.0:9000
 
+- `calculateRhoT()` and `calculateRhoC()` are now exported and support all of cSEM's    `cSEMResults` classes.
+  
 - `calculateGoF()` is now exported and supports all of cSEM's `cSEMResults` classes.
   Breaking: the `.only_common_factors` argument has been removed.